requirements.txt
langchain 
pandas 
faiss-cpu 
flagai